// DXC_Frontend/src/api/models/organizationWithDepartmentsDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { DepartmentDto } from './departmentDto'

export interface OrganizationWithDepartmentsDto {
  publicId?: string
  /** @nullable */
  code: string | null
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  departmentCount?: number
  /** @nullable */
  departments?: DepartmentDto[] | null
}